backend/requirements.txt
# Backend Requirements
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
gunicorn>=21.0.0
whitenoise>=6.6.0
